backend/requirements.txt
# ── Core Framework ──────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic[email]>=2.9.0
pydantic-settings>=2.0

# ── LLM Providers ───────────────────────────
openai>=1.0
google-generativeai>=0.7.0,<0.8.0

# ── Auth & Database ─────────────────────────
sqlalchemy>=2.0
psycopg2-binary>=2.9
psycopg[binary]
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# ── Storage & Cloud ─────────────────────────
boto3>=1.34

# ── LangChain Ecosystem ─────────────────────
# (Pinned to 0.2.x for maximum stability)
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-google-genai==1.0.10
langchain-huggingface==0.0.3
langchain-community==0.2.16
langchain-chroma==0.1.4
langchain-text-splitters==0.2.4

# ── Vector Store ────────────────────────────
chromadb==0.5.23
sentence-transformers>=2.2

# ── Document Parsing ────────────────────────
pypdf>=4.0.0
python-docx>=1.0

# ── Testing & Utils ─────────────────────────
httpx>=0.27
pytest>=7.0
pytest-asyncio>=0.23